added dtype

+++ model/dataset_loader.py
@@ -164,8 +164,8 @@ class GrooveMidiDataset(Dataset):
             json.dump(parameters, f)
 
         # convert inputs and outputs to torch tensors
-        self.processed_inputs = torch.Tensor(self.processed_inputs, device=device)
-        self.processed_outputs = torch.Tensor(self.processed_outputs, device=device)
+        self.processed_inputs = torch.Tensor(self.processed_inputs, device=device).to(torch.float32)
+        self.processed_outputs = torch.Tensor(self.processed_outputs, device=device).to(torch.float32)
 
     def get_hvo_sequence(self, idx):
         hvo_idx = self.hvo_index[idx]
